arch: changing arch
