+++ src/components/Checkbox/index.tsx
@@ -1,10 +1,6 @@
 import { FunctionComponent } from 'react'
 
-import {
-  CheckboxContainer,
-  HiddenCheckbox,
-  StyledCheckbox,
-} from './checkbox.styles'
+import { CheckboxContainer, HiddenCheckbox, StyledCheckbox } from './styles'
 import { Check } from '@phosphor-icons/react'
 
 export interface ICheckboxProps {
