+++ src/components/Counter/styles.ts
