+++ src/components/Input/index.tsx
@@ -1,4 +1,4 @@
-import { StyledInput } from './input.styles'
+import { StyledInput } from './styles'
 import { FC, InputHTMLAttributes } from 'react'
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {}
