+++ src/components/TaskManager/index.tsx
@@ -1,37 +1,45 @@
-import { ChangeEventHandler, FormEventHandler, useState } from 'react'
-import {faker} from '@faker-js/faker'
+import {
+  ChangeEventHandler,
+  FormEventHandler,
+  useCallback,
+  useState,
+} from 'react'
+import { faker } from '@faker-js/faker'
 
-import { Input } from '../Input/Input'
-import { Tasks } from '../Tasks/Tasks'
+import { Input } from '../Input'
+import { Tasks } from '../Tasks'
 
 import Plus from '../../assets/plus.svg'
 import { CreateTaskButton, TaskManagerContainer } from './styles'
 
 export interface ITask {
-    id: string
-    description: string
-    isCompleted: boolean
-  }
+  id: string
+  description: string
+  isCompleted: boolean
+}
 
 export const TaskManager = () => {
-    const [tasks, setTasks] = useState<ITask[]>([])
-    const [newTask, setNewTask] = useState('')
-  
-    const handleNewTask: ChangeEventHandler<HTMLInputElement> = (event) =>
-      setNewTask(event.target.value)
-  
-    const handleSubmitNewTask: FormEventHandler<HTMLFormElement> = (event) => {
+  const [tasks, setTasks] = useState<ITask[]>([])
+  const [newTask, setNewTask] = useState('')
+
+  const handleNewTask: ChangeEventHandler<HTMLInputElement> = (event) =>
+    setNewTask(event.target.value)
+
+  const handleSubmitNewTask: FormEventHandler<HTMLFormElement> = useCallback(
+    (event) => {
       event.preventDefault()
-  
+
       if (newTask) {
         setTasks((previousTasks) => [
           ...previousTasks,
           { description: newTask, id: faker.string.uuid(), isCompleted: false },
         ])
-  
+
         setNewTask('')
       }
-    }
+    },
+    [newTask],
+  )
 
   return (
     <TaskManagerContainer>
@@ -41,7 +49,7 @@ export const TaskManager = () => {
           placeholder="Adicione uma nova tarefa"
           value={newTask}
         />
-        <CreateTaskButton disabled={!newTask} >
+        <CreateTaskButton disabled={!newTask}>
           Criar
           <img src={Plus} alt="" title="Adicicionar to do" />
         </CreateTaskButton>
